% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{cleanrColours}
\alias{cleanrColours}
\title{cleanr theme colours}
\format{
An object of class \code{character} of length 6.
}
\usage{
cleanrColours
}
\description{
This is a vector with 6 colours to be used in palettes and other visual
elements.
}
\seealso{
\code{\link{cleanrPalette}}
}
\keyword{datasets}
