% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{cleanrPalette}
\alias{cleanrPalette}
\title{cleanr theme palette}
\usage{
cleanrPalette(n = 5)
}
\arguments{
\item{n}{the number of colours to output}
}
\description{
This function outputs n colours from the cleanr ggplot2 theme palette
}
\seealso{
\code{\link{cleanrCols}}
}
