% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cleanr.R
\name{theme_cleanr}
\alias{theme_cleanr}
\title{cleanr theme}
\usage{
theme_cleanr(type_family = "Avenir", type_size = 12)
}
\arguments{
\item{family}{Change the font family. Defaults to Fira Sans}
}
\value{
ggplot theme
}
\description{
This theme uses Mozilla's Fira Sans as its font.
Save to pdf using \code{cleanrSave()}.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg*0.43, y = wt*0.4535924, colour = factor(cyl))) +
  geom_point(size = 2) +
  labs(title = "Car weight vs efficiency",
       subtitle = "Using sensible metrics",
       x = "Efficiency (km/l)",
       y = "Weight (1000 kg)",
       colour = "Cylinders") +
  theme_cleanr() +
  scale_colour_cleanr()

}
\seealso{
\code{\link{cleanrSave}}
}
